+++ src/starccato_jax/vae/core/__init__.py
@@ -1,3 +1,3 @@
 from .io import load_model
-from .model import VAE, ModelData, encode, generate, reconstruct
+from .model import VAE, ModelData, encode, encode_mean, generate, reconstruct
 from .trainer import train_vae